\name{coef.arima.rob}
\alias{coef.arima.rob}
\title{
  Use coef on an arima.rob Object
}
\description{
This is a method for the function
\code{coef()} for objects inheriting from
class \code{"arima.rob"}.
See \code{coef} or
\code{coef.default} for the general behavior of this
function and for the interpretation of object.
}
\usage{
\method{coef}{arima.rob}(object, \dots)
}
\arguments{
\item{object}{
an object of class \code{"arima.rob"}.
}
\item{\dots}{
extra arguments passed to or from other methods.
The coef method here ignore these arguments.
}
}
\keyword{ts}
\keyword{robust}
\keyword{regression}
% docclass is function
% Converted by mySd2Rd version 47442.
