\name{frip.dat}
\alias{frip.dat}
\docType{data}
\title{
  Monthly Industrial Production of France
}
\format{A \code{"timeSeries"} vector}
\usage{
frip.dat
}
\description{
A \code{"timeSeries"} vector containing the monthly industrial production
of France from January 1960 to December 1989.
}
\keyword{sysdata}
