\name{import.dat}
\alias{import.dat}
\docType{data}
\title{
  Monthly Imports and Import Taxes of Argentina
}
\format{
A data frame with 96 observations and 2 columns:
\describe{
\item{taxes}{
monthly import taxes of Argentina.
}
\item{import}{
monthly imports of Argentina.
}
}}
\usage{
import.dat
}
\description{
The \code{import.dat} data frame has 96 rows and 2 columns.
The sample runs from January 1983 to December 1990.
}
\keyword{sysdata}
