\name{newtaxes.dat}
\alias{newtaxes.dat}
\docType{data}
\title{
  Monthly Import Taxes of Argentina
}
\format{
A data frame with 10 rows and 1 column:
\describe{
\item{taxes}{
monthly import taxes of Argentina.
}
}}
\usage{
newtaxes.dat
}
\description{
The \code{newtaxes.dat} data frame has 10 rows and 1 column.
The sample runs from January to October 1992.
}
\keyword{sysdata}
