\name{print.outliers}
\alias{print.outliers}
\title{
  Use print on an outliers Object
}
\description{
This is a method for the function \code{print}
for objects inheriting from class \code{"outliers"}.
See \code{print} or
\code{print.default} for the general behavior
of this function and for the interpretation of
\code{x}.
}
\usage{
\method{print}{outliers}(x, digits = 4, \dots)
}
\arguments{
\item{x}{
an object of class \code{"outliers"}.
}
\item{digits}{
the number of digits to display.
}
\item{\dots}{
extra arguments passed to or from other methods.
}
}
\seealso{
\code{\link{outliers}},
\code{\link{print}}.
}
\keyword{print}
% docclass is function
% Converted by mySd2Rd version 47442.
