\name{print.summary.arima.rob}
\alias{print.summary.arima.rob}
\title{
  Use print on a summary.arima.rob Object
}
\description{
This is a method for the function \code{print}
for objects inheriting
from class \code{"summary.arima.rob"}.
See \code{print} or
\code{print.default} for the general
behavior of this function and for the interpretation of
\code{x}.
}
\usage{
\method{print}{summary.arima.rob}(x, digits = 4, \dots)
}
\arguments{
\item{x}{
an object of class \code{"summary.arima.rob"}.
}
\item{digits}{
the number of digits to display.
}
\item{\dots}{
extra arguments passed to or from other methods.
}
}
\value{
invisibly returns \code{x}.
Typically called for the printing side effect.
}
\seealso{
\code{\link{print}}
\code{\link{summary.arima.rob}}.
}
\keyword{print}
