\name{print.summary.outliers}
\alias{print.summary.outliers}
\title{
  Use print on a summary.outliers Object
}
\description{
This is a method for the function \code{print()}
for objects inheriting from
class \code{"summary.outliers"}.
See \code{print} or
\code{print.default} for the general
behavior of this function and for the interpretation of
\code{x}.
}
\usage{
\method{print}{summary.outliers}(x, digits = 4, \dots)
}
\arguments{
\item{x}{
an object of class \code{"summary.outliers"}.
}
\item{digits}{
the number of digits to display.
}
\item{\dots}{
extra arguments passed to or from other methods.
}
}
\value{
invisibly returns \code{x}.
Typically called for the printing side effect.
}
\seealso{
\code{\link{print}},
\code{\link{summary.outliers}}.
}
\keyword{print}
