\name{print.arima.rob}
\alias{print.arima.rob}
\title{
  Use print on an arima.rob Object
}
\description{
This is a method for the function \code{print()}
for objects inheriting from
class \code{"arima.rob"}.
See \code{print} or
\code{print.default} for the general behavior
of this function and for the interpretation of
\code{x}.
}
\usage{
\method{print}{arima.rob}(x, digits = 4, \dots)
}
\arguments{
\item{x}{
an object of class \code{"arima.rob"}.
}
\item{digits}{
the number of digits to display.
}
\item{\dots}{
extra arguments passed to or from other methods.
}
}
\seealso{
\code{\link{print}}.
}
\keyword{print}
% docclass is function
% Converted by mySd2Rd version 47442.
